shark-indexer/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
asyncpg>=0.29.0
aiohttp>=3.9.1
pydantic>=2.5.2
python-dotenv>=1.0.0
alembic>=1.13.0
prometheus-client>=0.19.0
pytest>=7.4.3
pytest-asyncio>=0.23.2
httpx>=0.25.2
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.1
redis>=5.0.1
structlog>=24.1.0
requests>=2.31.0 